+++ apps/backoffice-v2/src/common/octet-to-file-type/octet-to-file-type.ts
@@ -1,0 +1,2 @@
+export const octetToFileType = (base64: string, fileType: string) =>
+  base64?.replace(/application\/octet-stream/gi, fileType);