+++ apps/backoffice-v2/src/common/utils/is-valid-url/is-valid-url.ts
@@ -1,0 +1,10 @@
+import { z } from 'zod';
+
+/**
+ * @description Checks if a passed string is a URL string.
+ * @param value
+ * @param
+ */
+export const isValidUrl = (value: unknown): value is string => {
+  return z.string().url().safeParse(value).success;
+};