+++ packages/common/src/types/any-record/index.ts
@@ -1,1 +1,0 @@
-export { AnyRecord } from './any-record';