+++ services/workflows-service/src/workflow/dtos/workflow-event-input.ts
@@ -1,6 +1,5 @@
 import { ApiProperty } from '@nestjs/swagger';
-import { IsOptional, IsString } from 'class-validator';
-import type { ResubmissionReason } from '@/workflow/workflow.service';
+import { IsString } from 'class-validator';
 
 export class WorkflowEventInput {
   @ApiProperty({
@@ -9,26 +8,4 @@ export class WorkflowEventInput {
   })
   @IsString()
   name!: string;
-
-  /**
-   * Which document requires re-submission - only required for re-submission
-   */
-  @ApiProperty({
-    required: false,
-    type: String,
-  })
-  @IsOptional()
-  @IsString()
-  document?: string;
-
-  /**
-   * Only required for re-submission
-   */
-  @ApiProperty({
-    required: false,
-    type: String,
-  })
-  @IsOptional()
-  @IsString()
-  resubmissionReason?: keyof typeof ResubmissionReason;
 }